feat(accounts): PER-221 — Edit + Reconcile/Update-value on the account detail page
Surface the two account mutations users most want while looking AT an account —
Edit metadata and Reconcile (cash) / Update value (tracked) — directly on the
per-account detail route, instead of forcing a trip back to the list.

To avoid a second copy, the dialogs are extracted verbatim from
accounts.index.tsx into shared components that both routes now consume:
- src/components/blocks/account-form-dialog.tsx (AccountFormDialog + AccountFormState)
- src/components/blocks/valuation-action-dialog.tsx (ValuationActionDialog)

ACCOUNT_TYPE_LABEL moved to its canonical home @/lib/accounts (re-exported from
-account-card for existing importers) so a shared component no longer imports
from a route file. The valuation dialog now also shows the PER-217 "Reserved"
cell when a reserve is set. On save, the detail route refetches the account +
transaction collections so the hero/KPIs/trend update immediately.

Tests: account-detail e2e extended to open Edit (prefilled), open Reconcile
(asserts the Reserved cell), reconcile to a new balance, and confirm the hero +
safe-to-spend update. Full list-route dialog + valuation e2e set re-run green
(7 passed) to prove the extraction didn't regress the list.
